Use iterable cache for fetch & synchronizing users
