+++ src/okta/types/cache.ts
@@ -1,0 +1,16 @@
+import { OktaFactor, OktaUser, OktaUserGroup } from ".";
+
+export interface OktaUserCacheData {
+  user: OktaUser;
+  factors: OktaFactor[];
+  userGroups: OktaUserGroup[];
+}
+
+export interface OktaUserCacheEntry {
+  key: string;
+  data?: OktaUserCacheData;
+}
+
+export interface OktaCacheState {
+  fetchCompleted: boolean;
+}
