+++ src/okta/types/index.ts
@@ -1,5 +1,6 @@
 export * from "./applications";
 export * from "./client";
+export * from "./cache";
 export * from "./devices";
 export * from "./groups";
 export * from "./users";
